+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/DependencyManager.groovy
@@ -15,6 +15,10 @@ import life.qbic.business.samples.count.CountSamplesOutput
 import life.qbic.business.samples.download.DownloadSamples
 import life.qbic.business.samples.download.DownloadSamplesDataSource
 import life.qbic.business.samples.download.DownloadSamplesOutput
+import life.qbic.business.samples.info.GetSamplesInfo
+import life.qbic.business.samples.info.GetSamplesInfoDataSource
+import life.qbic.business.samples.info.GetSamplesInfoInput
+import life.qbic.business.samples.info.GetSamplesInfoOutput
 import life.qbic.datamodel.dtos.portal.PortalUser
 import life.qbic.datamodel.dtos.projectmanagement.Project
 import life.qbic.portal.sampletracking.communication.notification.MessageBroker
@@ -28,6 +32,8 @@ import life.qbic.portal.sampletracking.components.projectoverview.download.Downl
 import life.qbic.portal.sampletracking.components.projectoverview.download.ManifestPresenter
 import life.qbic.portal.sampletracking.components.projectoverview.subscribe.SubscribeProjectController
 import life.qbic.portal.sampletracking.components.projectoverview.subscribe.SubscribeProjectPresenter
+import life.qbic.portal.sampletracking.components.projectoverview.samplelist.FailedQCSamplesView
+import life.qbic.portal.sampletracking.components.projectoverview.samplelist.ProjectOverviewController
 import life.qbic.portal.sampletracking.datasources.Credentials
 import life.qbic.portal.sampletracking.datasources.OpenBisConnector
 import life.qbic.portal.sampletracking.datasources.database.DatabaseSession
@@ -58,6 +64,7 @@ class DependencyManager {
 
     private LoadProjectsDataSource loadProjectsDataSource
     private CountSamplesDataSource countSamplesDataSource
+    private GetSamplesInfoDataSource getSamplesInfoDataSource
     private DownloadSamplesDataSource downloadSamplesDataSource
     private SubscriptionDataSource subscriptionDataSource
 
@@ -110,6 +117,7 @@ class DependencyManager {
         loadProjectsDataSource = openBisConnector
 
         subscriptionDataSource = new SubscriptionsDbConnector(DatabaseSession.getInstance())
+        getSamplesInfoDataSource = openBisConnector
     }
 
     /**
@@ -136,13 +144,22 @@ class DependencyManager {
     private ProjectOverviewView createProjectOverviewView() {
         Subscriber currentUser = new Subscriber(portalUser.firstName, portalUser.lastName, portalUser.emailAddress)
         ProjectOverviewViewModel viewModel = new ProjectOverviewViewModel(projectResourceService, statusCountService, currentUser)
-        DownloadProjectController downloadController = setupDownloadProjectUsecase(viewModel)
         SubscribeProjectController subscribeProjectController = setupSubscribeProjectUseCase()
-        ProjectOverviewView view =  new ProjectOverviewView(notificationService, viewModel, downloadController, subscribeProjectController)
+        DownloadProjectController downloadController = setupDownloadProjectUseCase(viewModel)
+
+        FailedQCSamplesView failedQCSamplesView = new FailedQCSamplesView(notificationService)
+        ProjectOverviewController projectOverviewController = setupFailedQCUseCase(failedQCSamplesView.getPresenter())
+
+        ProjectOverviewView view =  new ProjectOverviewView(notificationService, viewModel, downloadController, failedQCSamplesView, projectOverviewController, subscribeProjectController)
         return view
     }
-    
-    private DownloadProjectController setupDownloadProjectUsecase(ProjectOverviewViewModel viewModel) {
+
+    private ProjectOverviewController setupFailedQCUseCase(GetSamplesInfoOutput output){
+        GetSamplesInfo getSamplesInfo = new GetSamplesInfo(downloadSamplesDataSource,getSamplesInfoDataSource, output)
+        return new ProjectOverviewController(getSamplesInfo)
+    }
+
+    private DownloadProjectController setupDownloadProjectUseCase(ProjectOverviewViewModel viewModel) {
         DownloadSamplesOutput output = new ManifestPresenter(notificationService, viewModel)
         DownloadSamples downloadSamples = new DownloadSamples(downloadSamplesDataSource, output)
         
@@ -178,6 +195,7 @@ class DependencyManager {
             countSamples.countReceivedSamples(it)
             countSamples.countQcFailedSamples(it)
             countSamples.countAvailableDataSamples(it)
+            countSamples.countLibraryPrepFinishedSamples(it)
         }
     }